% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Manually add an identification from e.g. MS/MS}
\usage{
add_id(pepmatched, ID_vector, name_vector, pepseq_vector, pepmass_vector)
}
\arguments{
\item{pepmatched}{Input object of class \code{pepmatched}, generated by the pepmatch() or pep.id() function of this package.}

\item{ID_vector}{A character vector with the IDs of the features that get the identifications. This is the value of the first column of the "matchlist". See vignette for an example. Can also be a single character string.}

\item{name_vector}{A character vector with the same length as the ID_vector, containing the names of the peptides.}

\item{pepseq_vector}{Optional character vector with the peptide sequences of the identifications to be added.}

\item{pepmass_vector}{Optional numeric vector with the precise masses of the identifications to be added.}
}
\value{
An object of class \code{pepmatched} with added identifications that can serve as input to downstream labelpepmatch functions.
}
\description{
Manually add identifications to a pepmatched object. This can both be used as an extension to mass match, or without mass matching.
}
\author{
Rik Verdonck
}
\seealso{
\code{\link{pep.id}}, \code{\link{pep.massmatch}}
}

