% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_lpm_db.R
\name{download_lpm_db}
\alias{download_lpm_db}
\title{Download a peptide database.}
\usage{
download_lpm_db(db = c("desertlocust", "celegans"))
}
\arguments{
\item{db}{Character. Database to be downloaded. Choice between "desertlocust", "celegans"}
}
\value{
A peptide database of standard format with columns \code{"name","MW"} and \code{"sequence"} and optional columns \code{"family"} and \code{"reference"}
}
\description{
Download a database of known peptides into your R-session.
}
\author{
Rik Verdonck
}

