% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locustdata.R
\docType{data}
\name{locustdata}
\alias{locustdata}
\title{Dataset with TMAB labelled locust peptides.}
\format{\preformatted{List of 2
 $ frame :'data.frame':	1425 obs. of  26 variables:
  ..$ id        : num [1:1425] 36286 36285 36044 36116 36274 ...
  ..$ z         : num [1:1425] 1 1 1 1 1 1 1 1 2 4 ...
  ..$ mz_1      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_2      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_3      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_4      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_5      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_6      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_7      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ mz_8      : num [1:1425] 1868 580 1160 934 361 ...
  ..$ Quantity_1: num [1:1425] 3252 68516 324316 279604 517488 ...
  ..$ Quantity_2: num [1:1425] 8164 67744 382124 382000 801112 ...
  ..$ Quantity_3: num [1:1425] 5696 140180 434388 428184 692692 ...
  ..$ Quantity_4: num [1:1425] 7056 115140 545744 478340 808132 ...
  ..$ Quantity_5: num [1:1425] 9080 168324 566776 564280 973372 ...
  ..$ Quantity_6: num [1:1425] 4444 65364 637096 554152 987188 ...
  ..$ Quantity_7: num [1:1425] 3212 87568 642764 461484 992424 ...
  ..$ Quantity_8: num [1:1425] 3572 103852 673296 467448 1030116 ...
  ..$ Ret_1     : num [1:1425] 33.6 30.7 32.7 33.6 32 ...
  ..$ Ret_2     : num [1:1425] 33.7 30.8 32.8 33.7 32.1 ...
  ..$ Ret_3     : num [1:1425] 33.5 30.7 32.5 33.5 31.9 ...
  ..$ Ret_4     : num [1:1425] 33.7 31 32.9 33.7 32.2 ...
  ..$ Ret_5     : num [1:1425] 33.8 31 32.9 33.8 32.3 ...
  ..$ Ret_6     : num [1:1425] 33.5 30.7 32.6 33.5 31.9 ...
  ..$ Ret_7     : num [1:1425] 33.5 30.8 32.7 33.5 32 ...
  ..$ Ret_8     : num [1:1425] 33.5 30.8 32.6 33.5 32 ...
 $ design:'data.frame':	8 obs. of  5 variables:
  ..$ RunName       : Factor w/ 8 levels "A","B","C","D",..: 1 2 3 4 5 6 7 8
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ LightCondition: Factor w/ 2 levels "greg","sol": 2 2 1 1 2 2 1 1
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ HeavyCondition: Factor w/ 2 levels "greg","sol": 1 1 2 2 1 1 2 2
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ Direction     : Factor w/ 2 levels "F","R": 1 1 2 2 1 1 2 2
  ..$ FileName      : Factor w/ 1 level "exported_featuredata_schistoTMAB2011_2015.csv": 1 1 1 1 1 1 1 1
 - attr(*, "class")= chr "lpm_input"
}}
\usage{
locustdata
}
\description{
Dataset with TMAB labelled locust peptides.
}
\details{
These are peptides from the corpora cardiaca of adult desert locusts (\emph{Schistocerca gregaria}) labelled with light (D0) and heavy (D9) TMAB. The conditions are sol and greg: solitarious and gregarious animals. Check the design matrix \code{locustdata$design} to have an overview of the labeling. The quantity used here is the peak intensity as determined by Progenesis LC-MS.
}
\author{
Rik Verdonck
}
\references{
Will be published soon.
}
\keyword{data}
\keyword{example}
\keyword{locust}
\keyword{schistocerca}

