% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_heatmap.R
\name{lpm_heatmap}
\alias{lpm_heatmap}
\title{Plot a heatmap.}
\usage{
lpm_heatmap(x, contrasts = F, prawcutoff = 0.05, padjcutoff = 0.05,
  FCcutoff = 1.25, main = "heatmap")
}
\arguments{
\item{x}{An object of class \code{lpm_linearmodel}}

\item{contrasts}{Logical. Should the heatmap be on the contrasts, or on the residuals?}

\item{prawcutoff}{Numeric. Cutoff for maximal raw p-value for features to be retained in the heatmap.}

\item{padjcutoff}{Numeric. Cutoff for maximal adjusted p-value for features to be retained in the heatmap.}

\item{FCcutoff}{Numeric. Cutoff for minimal log2 fold change.}

\item{main}{Character. Title of the heatmap.}
}
\description{
Make a heatmap from an \code{lpm_linearmodel} object. Uses the residuals of the null model of the lpm_linearmodel. This is a linear model with label as a main effect, without accounting for treatment. The lpm_heatmap function can be applied both on the raw residuals or on the contrasts within one peak pair.
}
\author{
Rik Verdonck
}

