% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_linearmodel.R
\name{lpm_linearmodel}
\alias{lpm_linearmodel}
\title{Linear models for labelpepmatch.}
\usage{
lpm_linearmodel(statlist, method = "vanilla", p.adjust.method = "BH",
  cores = 1, logtransformed = T, verbose = F)
}
\arguments{
\item{statlist}{An object of class \code{lpm_statlist}}

\item{method}{Character. See details.}

\item{p.adjust.method}{The method you want to use for correction for multiple testing. Choose between "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr" or "none". For more information, see \code{\link{p.adjust}}}

\item{cores}{Interger. Number of cores that can be used on the computer for calculation. When >1, the packages foreach and doSNOW (windows) or doMC (linux) will be loaded.}

\item{logtransformed}{Logical. Are your data already transformed on a log2 scale?}

\item{verbose}{Logical. If TRUE, verbose output is generated during model estimation. Might be helpful when running computationally demanding models to monitor progress.}
}
\description{
This function takes a \code{lpm_statlist} object and runs a linear model on it. In this version of the package, two models are available. See details.
}
\details{
The vanilla method runs a separate mixed model on each feature, using label effect as a covariate and run as a random effect. Hence, it corrects for a label bias within each feature separately. In the output you will find a p-value for the label effect for each separate feature. The complexmixed model is a mixed model ran on all features at once, with label effect nested in run as a covariate. This method is extremely powerful, but calculation times rise quickly, and hence it is only possible to use on a limited number of features (e.g. only mass matched features, only highest quantities etc.). The time complexity is estimated to be quasipolynomial nlog(n), and it is advised not to use this method for more than 50 features.
}
\author{
Rik Verdonck & Wouter De Haes
}

