% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_make.RGList.R
\name{lpm_make.RGList}
\alias{lpm_make.RGList}
\title{Bridge between labelpepmatch and limma.}
\usage{
lpm_make.RGList(x)
}
\arguments{
\item{x}{An object of class \code{lpm_statlist}.}
}
\description{
Turn a labelpepmatch \code{pepmatched} or \code{lpm_statlist} object in a limma RGList object.
}
\author{
Rik Verdonck & Tom Wenseleers
}

