% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_mockdata.R
\name{lpm_mockdata}
\alias{lpm_mockdata}
\title{Generate mock data.}
\usage{
lpm_mockdata(input, masslevel = 100, retlevel = 3,
  elutionunit = "minutes", graphics = F)
}
\arguments{
\item{input}{An object of class \code{lpm_input}}

\item{masslevel}{Numeric. Level at which the mass values should be shopped up.}

\item{retlevel}{Numeric. Level at which the retention time values should be shopped up. Careful: it is best to either use small values (like up to maximum 10\% of the range of retention times) for a tight coupling between m/z and retention time, or a value that is larger than the maximum retention time for a loose coupling between m/z and retention time. Intermediate values will chop up your data in chunks. Have a look at it with the graphics parameter on.}

\item{elutionunit}{Character. Sketchy. Default is "minutes", alternative is "seconds".}

\item{graphics}{Output a plot of the mock database versus the real database.}
}
\value{
An object of class \code{lpm_input}. The values for m/z and retention time are shuffeled. The quantity values are all set to 0.
}
\description{
Generates mock data on the basis of an \code{lpm_input} object. This function is integrated in \code{\link{pepmatch}} FDR estimation with its default parameters. The use outside of the \code{pepmatch} function is only advised if you want to explore this function, or if you want to use non-default parameters. The way this function works, is by chopping up values of masses (or actually of m/z) and retention times, randomizing them, and pasting them back together. This is important since mass/charge values are non-random (since masses are often close to integer values) and also the coupling of retention time with mass/charge is non-random. For example, mass values 1155.2 and 2456.2 will be reshuffled to 1456.2 and 2115.2 with a masslevel of 1000.
}
\author{
Rik Verdonck
}
\seealso{
\code{\link{pepmatch}}
}

