% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_refine.R
\name{lpm_refine}
\alias{lpm_refine}
\title{Refine \code{pepmatched} objects.}
\usage{
lpm_refine(pepmatched_object, labelthresh, elutionthresh, MWmin, MWmax,
  quantmin, quantmax, labelcountmin, labelcountmax, zmin, zmax,
  remove.more.labels.than.charges = F, remove.run = NULL,
  only.identified = FALSE)
}
\arguments{
\item{pepmatched_object}{Object of class \code{pepmatched} that you want to trim.}

\item{labelthresh}{Numeric. Threshold for molecular weight difference (in Dalton) between to peaks to differ from theoretical mass difference between two labelled peptides. Regardless of number of labels.}

\item{elutionthresh}{Numeric. Threshold for elution time difference between two peaks to be considered a peakpair.}

\item{MWmin}{Numeric. Minimal molecular weight of the peptide without labels.}

\item{MWmax}{Numeric. Maximal molecular weight of the peptide without labels.}

\item{quantmin}{Numeric vector of one or two elements. Minimal untransformed quantity for a feature to be retained. The highest value of the vector is the cutoff for the most abundant peak in a pair, the lowest for the least abundant. If you want to make sure you pick up extreme up-or downregulations, the vector should contain one zero or very small number. If you want to discart a peak pair once one of the peaks is below a quantity threshold, you can set the threshold with one minimal value (a vector with one element). This is equivalent to a vector with two equal elements.}

\item{quantmax}{Numeric. Maximal quantity (intensity or abundance). This is an AND function, so both peaks in a peak pair have to be above this quantity in order to be discarted.}

\item{labelcountmin}{Integer. Minimal number of labels.}

\item{labelcountmax}{Integer. Maximal number of labels.}

\item{zmin}{Integer. Minimal number of charges.}

\item{zmax}{Integer. Maximal number of charges.}

\item{remove.more.labels.than.charges}{Logical. Remove features that have more labels than charges. This is usually relevant since most labels are charged, so finding a peptide that has more labels than charges is often impossible.}

\item{remove.run}{Integer. A single value or a vector of runnumbers that should be discarted.}

\item{only.identified}{Logical. Only features that have been identified with mass match are retained.}
}
\description{
This function allows to make subselections of a \code{pepmatched} object with cutoffs for different parameters. In this way, the \code{pepmatch} function itself only has to be run once with not to strict parameters, and analysis can afterwards be refined. Notice that it is impossible to refine parameters to less strict values than the ones used to generate the pepmatched object.
}

