% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpm_volcanoplot.R
\name{lpm_volcanoplot}
\alias{lpm_volcanoplot}
\title{Plot a volcano plot.}
\usage{
lpm_volcanoplot(x, adjusted = T, plotlocator = F)
}
\arguments{
\item{x}{An object of class \code{lpm_linearmodel}}

\item{adjusted}{Logical. If TRUE, adjusted p-values are used. If FALSE, raw p-values are used.}

\item{plotlocator}{Logical. If TRUE, you can click on the plot and the identity of the nearest feature will be printed into the R-session. Click under the plot to return to your normal R-session.}
}
\description{
Make a volcanoplot from an \code{lpm_linearmodel} object. Here, every feature is plotted in the log2(fold change) * -log10(p-value) space. It is a very convenient way of visualizing multiple tests in one plot.
}
\author{
Rik Verdonck
}

