% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.statlist.R
\name{make.statlist}
\alias{make.statlist}
\title{Reorganize a \code{pepmatched} object into matrix format.}
\usage{
make.statlist(pepmatched_object, cutoff = 1, logtransform = T, labelthresh,
  elutionthresh, MWmin, MWmax, quantmin, quantmax, labelcountmin, labelcountmax,
  zmin, zmax, remove.more.labels.than.charges = F, remove.run = NULL,
  only.identified = FALSE)
}
\arguments{
\item{pepmatched_object}{Object of class \code{pepmatched} that you want to trim.}

\item{cutoff}{Numeric. Proportion of runs in which a feature should be found to be retained in statlists. Default is 1}

\item{logtransform}{Logical. Should the data be log2 transformed? Set to FALSE if data are already transformed in earlier step, or if you want to manually transform your data. Otherwise this is the best place to log2 transform your data.}

\item{labelthresh}{Numeric. Threshold for molecular weight difference (in Dalton) between to peaks to differ from theoretical mass difference between two labelled peptides. Regardless of number of labels.}

\item{elutionthresh}{Numeric. Threshold for elution time difference between two peaks to be considered a peakpair.}

\item{MWmin}{Numeric. Minimal molecular weight of the peptide without labels.}

\item{MWmax}{Numeric. Maximal molecular weight of the peptide without labels.}

\item{quantmin}{Numeric. Minimal quantity (intensity or abundance). This is an AND function, so both peaks in a peak pair have to be below this quantity in order to be discarted.}

\item{quantmax}{Numeric. Maximal quantity (intensity or abundance). This is an AND function, so both peaks in a peak pair have to be above this quantity in order to be discarted.}

\item{labelcountmin}{Integer. Minimal number of labels.}

\item{labelcountmax}{Integer. Maximal number of labels.}

\item{zmin}{Integer. Minimal number of charges.}

\item{zmax}{Integer. Maximal number of charges.}

\item{remove.more.labels.than.charges}{Logical. Remove features that have more labels than charges. This is usually relevant since most labels are charged, so finding a peptide that has more labels than charges is often impossible.}

\item{remove.run}{Integer. A single value or a vector of runnumbers that should be discarted.}

\item{only.identified}{Logical. Only features that have been identified with mass match are retained.}
}
\value{
An object of class \code{lpm_statlist}
}
\description{
This function takes an object of class \code{pepmatched} and transforms it to class \code{lpm_statlist}. This statlist is a format that is suited for statistical analysis. It also contains the same trimming parameters as the \code{\link{lpm_refine}} function.
}
\author{
Rik Verdonck
}
\seealso{
\code{\link{lpm_refine}} \code{\link{lpm_make.RGList}}
}

