% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pep.id.R
\name{pep.id}
\alias{pep.id}
\title{Mass match peak pairs from a pepmatched object.}
\usage{
pep.id(pepmatched, ID_thresh = 10, db, presetdb = NA, dbpath = NA,
  dbseparator = ",", dbheader = FALSE, masscorrection = FALSE,
  cores = 1, FDR = TRUE, iterations = 100, checkdb = F, graphics = F,
  verbose = FALSE)
}
\arguments{
\item{pepmatched}{Input object of class \code{pepmatched}, generated by the pepmatch() function of this package.}

\item{ID_thresh}{Numeric. Maximal allowed mass difference in ppm for identification.}

\item{db}{In case no preset database is chosen: a database (table) with the first 3 columns \code{"name","MW"} and \code{"sequence"} (as read in with the \code{\link{download_lpm_db}} function) The amino acid sequences have to obey rules when you want to use them for mass recalculation or generation of a decoy database. Amino acids have to be capitalized one letter codes. For details, see \code{\link{calculate_peptide_mass}}}

\item{presetdb}{A preset database. For the future (not ready yet)}

\item{dbpath}{Character. In case a local database is used. Should be the filepath of a table, separated by dbseparator, with first tree columns: name, mass in Dalton and sequence.}

\item{dbseparator}{Character. Column separator of database.}

\item{dbheader}{Logical. Does the database have a header? Default FALSE}

\item{masscorrection}{Logical. Should masses be corrected on the basis of identifications? Caution should be payed here, since this will only work with a sufficiently high number of real matches. This rather sketchy feature should be considered something interesting to have a look at, rather than a compulsory part of the pipeline. Always also run without mass correction and compare! Default is FALSE.}

\item{cores}{Interger. Number of cores that can be used on the computer for calculation. When >1, the packages foreach and doParallel will be used for multithreading.}

\item{FDR}{Logical. Test false discovery rate of peptide mass match identification by calculating a decoy database and look how many hits you get there. Uses \code{\link{generate_random_db}}.}

\item{iterations}{How many iterations of FDR should be ran? This might take some time for larger datasets.}

\item{checkdb}{Look if the masses and sequences in your database make sense. UNDER CONSTRUCTION!!!}

\item{graphics}{Only applies when FDR is TRUE.}

\item{verbose}{Logical. If TRUE, verbose output is generated during identifications.}
}
\value{
An object of class \code{pepmatched} with added mass matches that can be used for subsequent analysis using labelpepmatch functions.
}
\description{
Mass match peak pairs from a pepmatched object to known databases. Can call inbuilt databases, but you can also use your own local database.
}
\author{
Rik Verdonck & Gerben Menschaert
}
\seealso{
\code{\link{generate_random_db}}, \code{\link{pep.massmatch}}
}

