% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pepmatch.R
\name{pepmatch}
\alias{pepmatch}
\title{Identify peak pairs.}
\usage{
pepmatch(lpm_input, elutionthresh = 0.2, elutionunit = "minutes",
  labelthresh = 0.1, labelcountmax = 6, labellightmass, labelheavymass,
  label = "unspecified", minmolweight = 132, quantmin = 0, FDR = T,
  iterations = 10, cores = 1, verbose = FALSE)
}
\arguments{
\item{lpm_input}{Input object of class "LPM_input", generated by the read functions of labelpepmatch, or example data object present in package.}

\item{elutionthresh}{Numeric. Threshold for elution time difference between two peaks to be considered a peakpair. Default is 0.2 (minutes)}

\item{elutionunit}{Character. Default is "minutes", alternative is "seconds". This is sketchy, only just for FDR estimation.}

\item{labelthresh}{Numeric. Threshold for molecular weight difference (in Dalton) between to peaks to differ from theoretical mass difference between two labelled peptides. Regardless of number of labels. Default is 0.1.}

\item{labelcountmax}{Integer. Maximal number of labels allowed. Default is 6.}

\item{labellightmass}{Numeric. Mass of light label. Default is 128.1177, the mass of light TMAB}

\item{labelheavymass}{Numeric. Mass of heavy label. Default is 137.1728, the mass of heavy TMAB}

\item{label}{Character. Optional argument. If it is "TMAB", automatically the right values for light and heavy TMAB are used. Can be extended in the future with newer labels.}

\item{minmolweight}{Numeric. Minimal molecular weight for a peptide to be retained. Default is 132, the smallest possible peptide.}

\item{quantmin}{Numeric. Minimal untransformed quantity for a feature to be retained. Caution: this is an OR function, so if one of both peaks in a peakpair is above quantmin, the peakpair is retained.}

\item{FDR}{Logical. Generates mock data according to a restricted randomization procedure that produces mock data with a very comparable structure to the original data. Structure elements that are retained are the dispersion of features in the retention time - m/z space and the structure of decimals for m/z. This is important because decimals are non random in m/z data. For more information see \code{\link{lpm_mockdata}}.}

\item{iterations}{Integer: the number of iterations to estimate the FDR. Careful: for large datasets this can take a long time! A modest number (3 to 5) of iterations will already give a good indication of FDR.}

\item{cores}{Interger. Number of cores that can be used on the computer for calculation. When >1, the packages foreach and doParallel will be used for multithreading. Optimally, the number of cores is a multiple of the number of runs. Since every run makes up a single thread, it makes no sense to use more cores than the number of runs.}

\item{verbose}{Logical. Gives verbose output.}
}
\value{
An object of class \code{pepmatched} without mass matchings. This object can serve as an imput to the \code{pep.id} function, or in case of no mass matching, can go directly in \code{make.statlist}.
}
\description{
Identifies pairs of labelled peptides in an \code{lpm_input} object.
}
\details{
This is the central function of the labelpepmatch package. Candidate matching features are those features that have the same charge, that co-elute within a given time interval and that have a mass difference of an integer multiple (the number of labels) of the mass difference between two labels. However, only m/z values are given, and masses have to be calculated first (deconvolution). Note that in the case one or more labels are present, the deconvoluted mass will not be correct because an unknown number of charges originate from labels rather than from protons. However, this error is systematic and would by definition be the same for two features within a peak pair. In other words, the mass difference within a peak pair is still calculated correctly. The function follows a simple algorithm where features are first sorted based on their preliminarily deconvoluted masses and retention time. Every feature is then matched to a set of equally charged features that fall within the same retention time window. Once a peak pair is detected, the number of labels is inferred from the mass difference and the correct mass of the peptide is recalculated.
}
\author{
Rik Verdonck & Gerben Menschaert
}
\seealso{
\code{\link{lpm_mockdata}}
}

