% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QC_database.R
\name{QC_database}
\alias{QC_database}
\title{Quality check a database.}
\usage{
QC_database(db)
}
\arguments{
\item{db}{A database with the first 3 columns \code{"name","MW"} and \code{"sequence"} (as read in with the \code{\link{download_lpm_db}} function)}
}
\value{
A vector with differences between observed and theoretical masses. Plots the observed and theoretical masses onto each other.
}
\description{
Verify if masses and sequences in database match.
}
\author{
Rik Verdonck
}
\seealso{
\code{\link{calculate_peptide_mass}}
}

