% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.labelpepmatch.R
\name{read.labelpepmatch}
\alias{read.labelpepmatch}
\title{Read labelpepmatch input.}
\usage{
read.labelpepmatch(indata = file.choose(), designvector, sep = "\\t",
  dec = ".", samplenames = NA, verbose = FALSE)
}
\arguments{
\item{indata}{Character. A table with standard columns: id, z, mz * n, quantity * n and retention time * n}

\item{designvector}{Character. A vector with the design of the labelling. This is a vector that contains in this exact order: name of first condition, name of second condition, labelling order of first sample, labelling order of second sample, ... labelling order of last sample. The labelling order can either be \code{F} (forward, first condition is light), or \code{R} (reverse, first condition is heavy).}

\item{sep}{Character. The field separator of the data you read in. Default is \code{\\t}}

\item{dec}{Character. The decimal separator of the data you read in. Default is "."}

\item{samplenames}{Character. A vector with all the names of the samples.}

\item{verbose}{Logical. Gives verbose output.}
}
\value{
An object of class \code{lpm_input} that can serve as an imput to the \code{pepmatch} function.
}
\description{
Read in data in a standard format for labelpepmatch.
}
\author{
Rik Verdonck
}

