% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.progenesis.R
\name{read.progenesis}
\alias{read.progenesis}
\title{Read progenesis data.}
\usage{
read.progenesis(indata = file.choose(), designvector, samplenames = NA,
  sep = ",", dec = ".", intORabu = "int", chargename = "Charge",
  abundname = "Raw abundance", intensname = "Intensity",
  rettimename = "Sample retention time (min)", verbose = FALSE)
}
\arguments{
\item{indata}{Character. A table exported from progenesis LC-MS. Either reads a filepath, or reads the file directly if you are in the right folder. If no indata is specified, a prompt is presented.}

\item{designvector}{Character. A Character vector of length N + 2 with N = number of runs. A vector with the design of the labelling.  This is a vector that contains in this exact order: name of first condition, name of second condition, labelling order of first sample, labelling order of second sample, ... labelling order of last sample. The labelling order can either be \code{F} (forward, first condition is light), or \code{R} (reverse, first condition is heavy).}

\item{samplenames}{Character. A character vector with the names of your samples. The length should be the number of runs.}

\item{sep}{Character. The field separator of the data you read in. Default is ","}

\item{dec}{Character. The decimal separator of the data you read in. Default is "."}

\item{intORabu}{Character. Can be "int" or "abu", default "int". Will you work with the intensity (height of the peaks) or the abundance (surface under the peaks) of the data?}

\item{chargename}{Character. Name of the charge columns. Default is \code{Charge}}

\item{abundname}{Character. Name of abundance columns header. Default is \code{Raw abundance}}

\item{intensname}{Character. Name of intensity columns header. Default is \code{Intensity}}

\item{rettimename}{Character. Name of retention time columns header. Default is \code{Sample retention time (min)}}

\item{verbose}{Logical. Gives verbose output.}
}
\value{
An object of class \code{lpm_input} that that can be used for subsequent analysis using labelpepmatch functions. Feeds straigth in the \code{pepmatch} function.
}
\description{
Read in data generated with progenesis LC-MS. Use the standard output function in progenesis and write your file to a table or csv. Do not bother about extra statistics columns. Do not normalize data.
}
\author{
Rik Verdonck & Gerben Menschaert
}

