% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wormdata.R
\docType{data}
\name{wormdata}
\alias{wormdata}
\title{Dataset with TMAB labelled worm peptides.}
\format{\preformatted{List of 2
 $ frame :'data.frame':	5112 obs. of  26 variables:
  ..$ id        : num [1:5112] 52 77 98 109 328 336 392 406 408 412 ...
  ..$ z         : num [1:5112] 2 2 2 2 2 2 2 2 2 2 ...
  ..$ mz_1      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_2      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_3      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_4      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_5      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_6      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_7      : num [1:5112] 300 357 316 385 437 ...
  ..$ mz_8      : num [1:5112] 300 357 316 385 437 ...
  ..$ Quantity_1: num [1:5112] 8337600 2784898 12070 4357060 784775 ...
  ..$ Quantity_2: num [1:5112] 8493706 3092316 7645 2888636 693567 ...
  ..$ Quantity_3: num [1:5112] 4930026 2139606 1238075 3368120 672805 ...
  ..$ Quantity_4: num [1:5112] 3728062 3711008 4214 1648094 1156652 ...
  ..$ Quantity_5: num [1:5112] 960269 1872255 4620346 355810 904497 ...
  ..$ Quantity_6: num [1:5112] 1199628 2023318 1203373 576177 806132 ...
  ..$ Quantity_7: num [1:5112] 617110 1618121 3585914 239157 581701 ...
  ..$ Quantity_8: num [1:5112] 547073 1077971 3928426 289456 585546 ...
  ..$ Ret_1     : num [1:5112] 23.4 29 35.8 19.8 36.7 ...
  ..$ Ret_2     : num [1:5112] 23.2 29.1 35.7 19.8 36.5 ...
  ..$ Ret_3     : num [1:5112] 23.4 29.1 35.7 19.9 36.6 ...
  ..$ Ret_4     : num [1:5112] 23.5 29.4 35.9 20.1 36.8 ...
  ..$ Ret_5     : num [1:5112] 23.6 29.4 36 20.1 36.9 ...
  ..$ Ret_6     : num [1:5112] 23.5 29.4 35.9 20.1 36.8 ...
  ..$ Ret_7     : num [1:5112] 23.6 29.4 35.9 20.1 36.8 ...
  ..$ Ret_8     : num [1:5112] 23.5 29.4 35.9 20.1 36.8 ...
 $ design:'data.frame':	8 obs. of  5 variables:
  ..$ RunName       : Factor w/ 8 levels "A","B","C","D",..: 1 2 3 4 5 6 7 8
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ LightCondition: Factor w/ 2 levels "AEX5","N2": 2 2 2 2 1 1 1 1
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ HeavyCondition: Factor w/ 2 levels "AEX5","N2": 1 1 1 1 2 2 2 2
  .. ..- attr(*, "names")= chr [1:8] "1" "2" "3" "4" ...
  ..$ Direction     : Factor w/ 2 levels "F","R": 1 1 1 1 2 2 2 2
  ..$ FileName      : Factor w/ 1 level "AEX5_20150429_editedfeatures.csv": 1 1 1 1 1 1 1 1
 - attr(*, "class")= chr "lpm_input"
}}
\usage{
wormdata
}
\description{
Dataset with TMAB labelled worm peptides.
}
\details{
These are peptides from entire body homogenates of adult (\emph{Caenorhabditis elegans}) worms labelled with light (D0) and heavy (D9) TMAB. The conditions are N2 and AEX5: Bristol N2 wild type, and aex5 mutants. Check the design matrix \code{wormdata$design} to have an overview of the labeling.
}
\author{
Wouter Dehaes
}
\references{
Will be published soon.
}
\keyword{caenorhabditis}
\keyword{data}
\keyword{elegans}
\keyword{example}
\keyword{nematode}
\keyword{worm}

